change connect() to attach()
