+++ src/index.js
@@ -1,6 +1,6 @@
 // This is the main file that gets referenced by node
+import attach from './utils/attach'
 import config from './config'
-import connect from './connect'
 import install from './install'
 import middlewares from './middlewares'
 import transports from './transports'
@@ -8,8 +8,8 @@ import log from './log'
 import './console'
 
 module.exports = {
+  attach,
   config,
-  connect,
   install,
   middlewares,
   transports,
