+++ src/install.js
@@ -1,5 +1,4 @@
-import connect from './connect'
-import debug from './utils/debug'
+import attach from './utils/attach'
 
 /**
  * Installs the timber logger to route all stdout logs to the provided stream
@@ -7,11 +6,11 @@ import debug from './utils/debug'
  * @param {Stream} transport - the stream that all logs will go through
  */
 function install(transport) {
-  // console.info('transport')
   if (!transport) throw Error('No transport was provided.')
 
-  // connect our transport stream to stdout
-  connect(transport)
+  // attach our transport stream to stdout/stderr
+  attach(transport, process.stdout)
+  attach(transport, process.stderr)
 }
 
 export default install
